move deamon code into upload_logg
